Column: normalize numeric types
Remove deprecated ZEROFILL attribute for numeric data types, e.g. "int(9)" becames "int".

As of MySQL 8.0.17, the ZEROFILL attribute is deprecated for numeric data types and support for it will be removed in a future MySQL version. Consider using an alternative means of producing the effect of this attribute. For example, applications could use the LPAD() function to zero-pad numbers up to the desired width, or they could store the formatted numbers in CHAR columns.

-- https://dev.mysql.com/doc/refman/8.0/en/numeric-type-syntax.html